+++ src/codemagic/apple/app_store_connect/versioning/__init__.py
@@ -1,3 +1,4 @@
 from .app_store_version_submissions import AppStoreVersionSubmissions
 from .app_store_versions import AppStoreVersions
+from .beta_app_review_submissions import BetaAppReviewSubmissions
 from .pre_release_versions import PreReleaseVersions