+++ src/codemagic/apple/resources/beta_app_review_submission.py
@@ -1,0 +1,21 @@
+from __future__ import annotations
+
+from dataclasses import dataclass
+
+from .enums import BetaReviewState
+from .resource import Relationship
+from .resource import Resource
+
+
+class BetaAppReviewSubmission(Resource):
+    """
+    https://developer.apple.com/documentation/appstoreconnectapi/betaappreviewsubmission
+    """
+
+    @dataclass
+    class Attributes(Resource.Attributes):
+        betaReviewState: BetaReviewState
+
+    @dataclass
+    class Relationships(Resource.Relationships):
+        build: Relationship