+++ src/codemagic/apple/resources/enums.py
@@ -30,6 +30,13 @@ class AppStoreState(_ResourceEnum):
     REPLACED_WITH_NEW_VERSION = 'REPLACED_WITH_NEW_VERSION'
 
 
+class BetaReviewState(_ResourceEnum):
+    APPROVED = 'APPROVED'
+    IN_REVIEW = 'IN_REVIEW'
+    REJECTED = 'REJECTED'
+    WAITING_FOR_REVIEW = 'WAITING_FOR_REVIEW'
+
+
 class BuildProcessingState(_ResourceEnum):
     PROCESSING = 'PROCESSING'
     FAILED = 'FAILED'
@@ -122,6 +129,11 @@ class CertificateType(_ResourceEnum):
             raise ValueError(f'Certificate type for profile type {profile_type} is unknown')
 
 
+class ContentRightsDeclaration(_ResourceEnum):
+    DOES_NOT_USE_THIRD_PARTY_CONTENT = 'DOES_NOT_USE_THIRD_PARTY_CONTENT'
+    USES_THIRD_PARTY_CONTENT = 'USES_THIRD_PARTY_CONTENT'
+
+
 class DeviceClass(_ResourceEnum):
     APPLE_TV = 'APPLE_TV'
     APPLE_WATCH = 'APPLE_WATCH'
@@ -190,8 +202,10 @@ class ReleaseType(_ResourceEnum):
 
 
 class ResourceType(_ResourceEnum):
+    APPS = 'apps'
     APP_STORE_VERSIONS = 'appStoreVersions'
     APP_STORE_VERSION_SUBMISSIONS = 'appStoreVersionSubmissions'
+    BETA_APP_REVIEW_SUBMISSIONS = 'betaAppReviewSubmissions'
     BUILDS = 'builds'
     BUNDLE_ID = 'bundleIds'
     BUNDLE_ID_CAPABILITIES = 'bundleIdCapabilities'