+++ src/codemagic/tools/_xcode_project/arguments.py
@@ -96,6 +96,16 @@ class XcodeProjectArgument(cli.Argument):
         ),
         argparse_kwargs={'required': False, 'action': 'store_true'},
     )
+    IPA_PATH = cli.ArgumentProperties(
+        key='ipa_path',
+        type=cli.CommonArgumentTypes.existing_path,
+        description='Path to iOS App Store Package file (.ipa)',
+    )
+    PKG_PATH = cli.ArgumentProperties(
+        key='pkg_path',
+        type=cli.CommonArgumentTypes.existing_path,
+        description='Path to MacOs Application Package file (.pkg)',
+    )
 
 
 class ExportIpaArgument(cli.Argument):