+++ tests/apple/app_store_connect/versioning/test_pre_release_versions.py
@@ -10,12 +10,11 @@ from tests.apple.app_store_connect.resource_manager_test_base import ResourceMan
 
 
 @pytest.mark.skipif(not os.environ.get('RUN_LIVE_API_TESTS'), reason='Live App Store Connect API access')
-class BuildsTest(ResourceManagerTestsBase):
+class PreReleaseVersionsTest(ResourceManagerTestsBase):
     def test_list(self):
         resource_filter = self.api_client.pre_release_versions.Filter(app=ResourceId('1453997552'))
         pre_release_versions, builds = self.api_client.pre_release_versions.list_with_include(
             Build, resource_filter=resource_filter)
-        builds = self.api_client.builds.list()
         assert len(builds) > 0
         for build in builds:
             assert isinstance(build, Build)